src/frontend detector/requirements.txt
ultralytics
opencv-python
numpy
fastapi
uvicorn
python-multipart
requests
tqdm
supabase
python-dotenv